% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative conditionings.R
\name{cond_prob_3}
\alias{cond_prob_3}
\title{Called by \link{mbd_loglik} if there is a conditioning != 0}
\usage{
cond_prob_3(pars, brts, cond, n_0 = 2, tips_interval = c(n_0 * (cond >
  0), Inf), lx = 30, debug_mode = FALSE)
}
\value{
the conditional probability
}
\description{
Called by \link{mbd_loglik} if there is a conditioning != 0
}
\author{
Giovanni Laudanno, Bart Haegeman
}
