% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative conditionings.R
\name{cond_prob_rhs1_3}
\alias{cond_prob_rhs1_3}
\title{Auxilary function for cond_prob}
\usage{
cond_prob_rhs1_3(qvec, lambda, mu, nu, k, nu_matrix, m1, m2, empty_qq, t)
}
\description{
Auxilary function for cond_prob
}
\author{
Giovanni Laudanno, Bart Haegeman
}
