% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(pars, n_0, age, lx, lx_min, lx_max, saveit, verbose)
}
\arguments{
\item{pars}{parameters}

\item{n_0}{starting species}

\item{age}{The crown age of the tree}

\item{lx}{dimension of the matrix}

\item{lx_min}{min lx}

\item{lx_max}{max lx}

\item{saveit}{do you want to save it?}

\item{verbose}{give more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno.
}
