% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_asymmetric_conditioning.R
\name{measure_asymmetric_conditioning}
\alias{measure_asymmetric_conditioning}
\title{Measure Asymmetric conditioning}
\usage{
measure_asymmetric_conditioning(pars, age = 10, lx_min = 20,
  lx_max = 50, saveit = TRUE)
}
\arguments{
\item{pars}{parameters}

\item{age}{The crown age of the tree}

\item{lx_min}{min lx}

\item{lx_max}{max lx}

\item{saveit}{do you want to save it?}
}
\value{
a dataframe
}
\description{
Measure Asymmetric conditioning
}
\author{
Giovanni Laudanno
}
