% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asymmetric_conditioning.R
\name{plot_asymmetric_conditioning}
\alias{plot_asymmetric_conditioning}
\title{Measure Asymmetric conditioning}
\usage{
plot_asymmetric_conditioning(pars, age = 10, saveit = TRUE)
}
\arguments{
\item{pars}{parameters}

\item{age}{The crown age of the tree}

\item{saveit}{do you want to save it?}
}
\value{
a dataframe
}
\description{
Measure Asymmetric conditioning
}
\author{
Giovanni Laudanno
}
