% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all}
\usage{
run_all(pars, age = 10, lx_min = 20, lx_max = 50, saveit = TRUE)
}
\arguments{
\item{pars}{parameters}

\item{age}{The crown age of the tree}

\item{lx_min}{min lx}

\item{lx_max}{max lx}

\item{saveit}{do you want to save it?}
}
\value{
List with meaningful measures for each parameter setting
}
\description{
Run all
}
\author{
Giovanni Laudanno
}
